# IrcParser: make public and speed up by another <=15%