--- a/TwitchLib.Client.Benchmark/IrcMessageHandler.cs
+++ b/TwitchLib.Client.Benchmark/IrcMessageHandler.cs
@@ -1,5 +1,5 @@
 using BenchmarkDotNet.Attributes;
-using TwitchLib.Client.Internal.Parsing;
+using TwitchLib.Client.Parsing;
 
 namespace TwitchLib.Client.Benchmark
 {