--- a/TwitchLib.Client.Benchmark/TwitchLib.Client.Benchmark.csproj
+++ b/TwitchLib.Client.Benchmark/TwitchLib.Client.Benchmark.csproj
@@ -10,7 +10,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.13.5" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.7" />
   </ItemGroup>
 
   <ItemGroup>